refactor(test): вынести функцию в отдельный файл
