+++ test-utils/change-visibility-state.ts
@@ -1,0 +1,7 @@
+export const changeVisibilityState = (state: VisibilityState): void => {
+  Object.defineProperty(global.document, 'visibilityState', {
+    value: state,
+    configurable: true,
+  });
+  global.document.dispatchEvent(new Event('visibilitychange'));
+};
